Working launcher

+++ addons/minecraft_launcher/utils.gd
@@ -32,31 +32,51 @@ static func download_file(request: Requests, url: String, path: String, sha1: St
 	# if something is missing, don't do it
 	if url == "" || path == "": return
 	
-	if not FileAccess.file_exists(path) or (Utils.check_sha1(path, sha1) and overwrite):
+	#if not FileAccess.file_exists(path) or (Utils.check_sha1(path, sha1) and overwrite):
+	if not FileAccess.file_exists(path) or overwrite:
 		var response := await request.do_file(url, path)
 		if response.result != Requests.Result.SUCCESS:
-			print("Error of type %s: code %s" % [response.result, response.code])
+			printerr("Error %s while downloading file %s from %s - code: %s" % [response.result, path, url, response.code])
 			
 	# check if file is correct
 	if sha1 != "" and not Utils.check_sha1(path, sha1):
 		DirAccess.remove_absolute(path)
 
 static func unzip_file(path: String, exclude_files: Array[String], delete_archive: bool) -> void:
-	var reader = ZIPReader.new()
+	var reader := ZIPReader.new()
 	var err = reader.open(path)
 	
+	if err != OK:
+		printerr("Error  %s while opening zip file %s" % [err, path])
+		return
 	var files = reader.get_files()
-	for file in files:
-		if file.get_file() in exclude_files:
+	
+	for zip_file in files:
+		if zip_file.get_file() in exclude_files:
 			continue
-			
-		reader.extract_file(file, true, path.get_base_dir().path_join(file))
+		
+		extract_file(reader, zip_file, path.get_base_dir())
 	
 	reader.close()
 	
 	if delete_archive:
 		DirAccess.remove_absolute(path)
 
+static func extract_file(reader: ZIPReader, inner_path: String, outer_dst: String):
+	if not reader.file_exists(inner_path):
+		return
+	
+	var outer_path: String = outer_dst.path_join(inner_path)
+	var outer_folder = ProjectSettings.globalize_path(outer_path.get_base_dir())
+	if not DirAccess.dir_exists_absolute(outer_folder):
+		DirAccess.make_dir_recursive_absolute(outer_folder)
+	
+	var extracted_file = FileAccess.open(outer_path, FileAccess.WRITE)
+	if extracted_file != null:
+		var bytes := reader.read_file(inner_path, true)
+		extracted_file.store_buffer(bytes)
+		extracted_file.close()
+		FileAccess.set_unix_permissions(outer_path, 511) #TODO: use custom build
 
 static func get_os_type() -> OS_TYPE:
 	var os_name = OS.get_name()
